requirements.txt
sqlalchemy
pandas
python-dotenv
hypothesis[pytest]
pytest
